+++ src/graphmap/postprocess_lcs.cc
@@ -334,8 +334,8 @@ int GraphMap::PostProcessRegionWithLCS_(ScoreRegistry* local_score, MappingData*
   }
 
   // Find the L1 parameters (median line and the confidence intervals).
-  float l_diff = read->get_sequence_length() * parameters->error_rate;
-  float maximum_allowed_deviation = l_diff * sqrt(2.0f) / 2.0f;
+  double l_diff = read->get_sequence_length() * parameters->error_rate;
+  double maximum_allowed_deviation = l_diff * sqrt(2.0f) / 2.0f;
   float sigma_L2 = 0.0f, confidence_L1 = 0.0f;
   int64_t k = 0, l = 0;
   // Actual L1 calculation.
@@ -425,7 +425,7 @@ int GraphMap::PostProcessRegionWithLCS_(ScoreRegistry* local_score, MappingData*
   l1_info.l1_confidence_abs = confidence_L1;
   l1_info.l1_std = sigma_L2;
   l1_info.l1_rough_start = l1_info.l1_k * 0 + l1_info.l1_lmin;
-  l1_info.l1_rough_end = l1_info.l1_k * read->get_sequence_length() + l1_info.l1_lmax;
+  l1_info.l1_rough_end = ((double) l1_info.l1_k * read->get_sequence_length()) + l1_info.l1_lmax;
   if (l1_info.l1_rough_start < index->get_reference_starting_pos()[local_score->get_region().reference_id])
     l1_info.l1_rough_start = index->get_reference_starting_pos()[local_score->get_region().reference_id];
   if (l1_info.l1_rough_end >= (index->get_reference_starting_pos()[local_score->get_region().reference_id] + index->get_reference_lengths()[local_score->get_region().reference_id]))
